requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1 ;python_version<"3.9"
cffi==1.15.1
cryptography==37.0.4
dj-database-url==1.0.0
Django==4.1
django-cors-headers==3.13.0
django-rest-knox==4.2.0
djangorestframework==3.13.1
gunicorn==20.1.0
psycopg2==2.9.3
pycparser==2.21
python-dotenv==0.20.0
pytz==2022.1
sqlparse==0.4.2
whitenoise==6.2.0